pyproject.toml
[tool.poetry]
name = "data-engineer"
version = "0.1.0"
description = ""
authors = ["Somitra <smtrgupta96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.8.6"
click = "^8.0.1"
pandas = "^1.2.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
